requirements.txt
web3
eth-account
bip-utils
requests
termcolor
python-dotenv